+++ src/main/java/com/bitbank/config/SecurityConfiguration.java
@@ -5,9 +5,13 @@ import java.util.List;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
@@ -22,13 +26,31 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
 	}
 
 	@Override
+	@Bean
+	public AuthenticationManager authenticationManagerBean() throws Exception {
+		return super.authenticationManagerBean();
+	}
+
+	@Override
 	protected void configure(HttpSecurity http) throws Exception {
 
 		// Enable CORS and disable CSRF
 		http = http.cors().and().csrf().disable();
-		http.authorizeRequests().mvcMatchers("/api/v1/ping").permitAll()
+		http.authorizeRequests().mvcMatchers("/api/v1/login").permitAll().mvcMatchers("/api/v1/ping").permitAll()
 				.mvcMatchers("/api/v1/bank-account-transactions").permitAll().anyRequest().authenticated();
+		
 	}
+	
+	// Fix There is no PasswordEncoder mapped for the id "null"
+	@Override
+    protected void configure(AuthenticationManagerBuilder auth) 
+      throws Exception {
+        PasswordEncoder encoder = PasswordEncoderFactories.createDelegatingPasswordEncoder();
+        auth.inMemoryAuthentication()
+          .withUser("username")
+          .password(encoder.encode("password"))
+          .roles("USER");
+    }
 
 	@Bean
 	public CorsFilter corsFilter() {